ai-service/requirements.txt
fastapi
uvicorn
python-multipart
requests
python-dotenv
boto3
langchain
langgraph
langchain-huggingface
huggingface_hub
faiss-cpu
sentence-transformers
